migration/requirements.txt
# Dependencies for MariaDB to PostgreSQL migration script

# MariaDB/MySQL connection
pymysql>=1.1.0,<2.0

# PostgreSQL connection with extended types support
psycopg2-binary>=2.9.9,<3.0

# Environment variables management
python-dotenv>=1.0.1,<2.0

# HTTP requests with retry
requests>=2.31.0,<3.0
urllib3>=2.0.0,<3.0

# Test dependencies
pytest>=7.0.0,<8.0
pytest-cov>=4.0.0,<5.0
pytest-mock>=3.10.0,<4.0
